refactor(cli): move the generic flag help into command-flag-help
Keeping formatFlagHelp in help.ts made the two modules import each other,
which the native code-quality audit rejects as a cycle.
