+++ src/cli/command-flag-help.ts
@@ -1,5 +1,3 @@
-import { formatFlagHelp } from './help'
-
 // Per-command overrides of the shared flag help; falls back to the generic text.
 export function formatCommandFlagHelp(flag: string, commandPath: string[]): string {
   const command = commandPath.join(' ')
@@ -111,4 +109,140 @@ export function formatCommandFlagHelp(flag: string, commandPath: string[]): stri
     return '--key <key>            Single key, e.g. Return, Escape, Tab, Left, or PageUp'
   }
   return formatFlagHelp(flag)
+}
+
+export function formatFlagHelp(flag: string): string {
+  const helpByFlag: Record<string, string> = {
+    agent: '--agent <id>          Launch a known TUI agent in the first terminal',
+    'base-branch': '--base-branch <ref>    Base branch/ref to create the worktree from',
+    command: '--command <text>       Command to run in the terminal on startup',
+    comment: '--comment <text>       Comment stored in Orca metadata',
+    cursor: '--cursor <n>           Line cursor from a previous read (returns only new output)',
+    action: '--action <name>       Secondary accessibility action name',
+    activate: '--activate             Reveal the new worktree in the Orca app',
+    app: '--app <app>            App name, bundle ID, or pid:N',
+    direction:
+      '--direction <dir>      Direction: up|down|left|right for scroll, horizontal|vertical for split',
+    'display-name': '--display-name <name>  Override the Orca display name',
+    'element-index': '--element-index <n>   Element index from get-app-state',
+    title: '--title <text>         Custom title for the terminal tab (omit to reset)',
+    enter: '--enter                Append Enter after sending text',
+    force:
+      '--force                Force worktree removal when supported; does not force branch deletion',
+    focus: '--focus                Reveal the created terminal session in Orca',
+    for: '--for exit|tui-idle    Wait condition to satisfy',
+    'from-element-index': '--from-element-index <n> Source element index from get-app-state',
+    'from-x': '--from-x <x>           Source window-local x coordinate',
+    'from-y': '--from-y <y>           Source window-local y coordinate',
+    help: '--help                 Show this help message',
+    'include-visual-layouts':
+      '--include-visual-layouts Include tab and pane topology in JSON output',
+    interrupt: '--interrupt            Send as an interrupt-style input when supported',
+    id: '--id <id>             Identifier for a target item or permission',
+    issue: '--issue <number|null>  Linked GitHub issue number',
+    'linear-issue':
+      '--linear-issue <id|url|null> Linked Linear issue identifier or URL; null clears on set',
+    json: '--json                 Emit machine-readable JSON',
+    key: '--key <key>            Key argument for this command',
+    limit: '--limit <n>            Maximum number of rows to return',
+    local: '--local                Target the current project instead of the global install',
+    skill: '--skill <name>         Bundled skill to act on; repeat for several',
+    mode: '--mode <mode>          Mode such as edit, diff, or both',
+    model: '--model <id>          Provider model id for a new agent launch',
+    effort: '--effort <level>      Reasoning effort for the selected model',
+    'mouse-button': '--mouse-button <btn>   Mouse button: left, right, or middle',
+    modifiers: '--modifiers <chord>  Modifier keys held only for this click',
+    name: '--name <name>          Name for the new worktree or automation',
+    'no-parent': '--no-parent            Force no parent lineage for unrelated work',
+    'no-screenshot': '--no-screenshot       Skip screenshot capture after the operation',
+    pages: '--pages <n>           Number of scroll pages',
+    'parent-worktree':
+      '--parent-worktree <selector> Parent worktree selector such as identity:<identity>, id:<repo-id>::<path>, branch:<branch>, issue:<number>, path:<path>, or active/current',
+    path: '--path <path>          Path argument for the command',
+    prompt: '--prompt <text>        Prompt text for agent-backed commands',
+    query: '--query <text>        Search text for matching refs',
+    ref: '--ref <ref>            Base ref to persist for the repo',
+    repo: '--repo <selector>      Repo selector such as id:<id>, name:<name>, or path:<path>',
+    'restore-window':
+      '--restore-window     Bring the target app/window forward before the operation',
+    session: '--session <id>        Snapshot namespace for a related computer-use workflow',
+    setup: '--setup run|skip|inherit Setup policy for repo-defined setup hooks',
+    terminal: '--terminal <handle>  Runtime-issued terminal handle',
+    text: '--text <text>          Text payload to send or type',
+    'text-stdin': '--text-stdin          Read text payload from stdin',
+    'task-id': '--task-id <id>        Task id to include in orchestration payload JSON',
+    'task-title': '--task-title <text>    Concise title for an orchestration task',
+    'dispatch-id': '--dispatch-id <id>    Dispatch id to include in orchestration payload JSON',
+    'files-modified': '--files-modified <csv> Comma-separated files for orchestration payload JSON',
+    'report-path': '--report-path <path>  Report path to include in orchestration payload JSON',
+    phase: '--phase <text>        Worker phase to include in orchestration payload JSON',
+    'timeout-ms': '--timeout-ms <ms>     Maximum wait time before timing out',
+    'to-element-index': '--to-element-index <n> Destination element index from get-app-state',
+    'to-x': '--to-x <x>             Destination window-local x coordinate',
+    'to-y': '--to-y <y>             Destination window-local y coordinate',
+    worktree:
+      '--worktree <selector>  Worktree selector such as identity:<identity>, id:<repo-id>::<path>, name:<displayName>, branch:<branch>, issue:<number>, path:<path>, or active/current',
+    workspace: '--workspace <selector> Existing worktree selector for automation runs',
+    'workspace-status':
+      '--workspace-status <id> Board status id (defaults: todo, in-progress, in-review, completed)',
+    staged: '--staged               Open staged source-control changes',
+    provider: '--provider <agent>     Agent id such as codex, claude, or gemini',
+    'source-context':
+      '--source-context <json|null> Explicit TaskSourceContext for automation task/provider data',
+    trigger: '--trigger <schedule>   Automation schedule preset, cron, or RRULE',
+    schedule: '--schedule <schedule>  Alias for --trigger',
+    time: '--time <HH:MM>        Time used with daily/weekdays/weekly presets',
+    day: '--day <0-6>           Day used with weekly preset, Sunday=0',
+    timezone: '--timezone <tz>       IANA timezone for the automation',
+    enabled: '--enabled              Enable the automation',
+    disabled: '--disabled             Disable the automation',
+    'reuse-session':
+      '--reuse-session        Reuse the previous live session for existing-workspace runs',
+    'fresh-session': '--fresh-session        Disable session reuse for future runs',
+    'workspace-mode': '--workspace-mode <mode> existing or new-per-run',
+    'missed-run-grace-minutes': '--missed-run-grace-minutes <n> Missed-run grace window',
+    'value-stdin': '--value-stdin         Read set-value payload from stdin',
+    'window-id': '--window-id <id>      Target a window id from list-windows',
+    'window-index': '--window-index <n>   Target a window index from list-windows',
+    // Browser automation flags
+    element: '--element <ref>        Element ref from snapshot (e.g. e3)',
+    url: '--url <url>            URL to navigate to',
+    value: '--value <text>         Value to fill or select',
+    input: '--input <text>         Text to type at current focus',
+    expression: '--expression <js>     JavaScript expression to evaluate',
+    amount: '--amount <pixels>      Scroll distance in pixels',
+    index: '--index <n>            Tab index to switch to',
+    page: '--page <id>            Stable browser page id from `orca tab list --json`',
+    profile: '--profile <id>        Browser profile id',
+    'show-profile': '--show-profile        Include tab profile in text output',
+    'no-ua-spoof': "--no-ua-spoof         Keep Electron's native user agent",
+    format: '--format <png|jpeg>    Screenshot image format'
+  }
+
+  if (flag === 'current') {
+    return '--current              Use the current Orca worktree linked Linear issue'
+  }
+  if (flag === 'comments') {
+    return '--comments             Include threaded Linear comments'
+  }
+  if (flag === 'children') {
+    return '--children             Include recursive child issues'
+  }
+  if (flag === 'depth') {
+    return '--depth <n>            Child issue depth for --children/--full'
+  }
+  if (flag === 'attachments') {
+    return '--attachments          Include attachment metadata and URLs'
+  }
+  if (flag === 'relations') {
+    return '--relations            Include blocking, related, and duplicate links'
+  }
+  if (flag === 'activity') {
+    return '--activity             Include issue field-change history'
+  }
+  if (flag === 'full') {
+    return '--full                 Include all supported V1 issue context within caps'
+  }
+
+  return helpByFlag[flag] ?? `--${flag}`
 }
